qty done, working on /by_username/:username

+++ src/domains/comment/service/comment.service.impl.ts
@@ -1,5 +1,5 @@
 import { CreateCommentInputDTO } from '../dto'
-import { PostDTO } from '@domains/post/dto'
+import { PostDTO, ExtendedPostDTO } from '@domains/post/dto'
 import { CommentRepository } from '../repository'
 import { CommentService } from '.'
 import { FollowerRepositoryImpl } from '@domains/follower/repository'
@@ -8,22 +8,27 @@ import { UserServiceImpl } from '@domains/user/service'
 import { UserRepositoryImpl } from '@domains/user/repository'
 import { validate } from 'class-validator'
 import { ForbiddenException, NotFoundException, db } from '@utils'
+import { CursorPagination } from '@types'
+import { PostRepositoryImpl } from '@domains/post/repository'
 
 const followerService = new FollowerServiceImpl(new FollowerRepositoryImpl(db))
 const userService = new UserServiceImpl(new UserRepositoryImpl(db))
+const PostRepository = new PostRepositoryImpl(db)
 
 export class CommentServiceImpl implements CommentService {
   constructor (private readonly repository: CommentRepository) {}
 
   async createComment (userId: string, data: CreateCommentInputDTO): Promise<PostDTO> {
     await validate(data)
+    await PostRepository.addQtyComments(data.parentId)
     return await this.repository.create(userId, data)
   }
 
   async deleteComment (userId: string, commentId: string): Promise<void> {
     const comment = await this.repository.getById(commentId)
-    if (!comment) throw new NotFoundException('comment')
+    if (!comment || !comment.parentId) throw new NotFoundException('comment')
     if (comment.authorId !== userId) throw new ForbiddenException()
+    await PostRepository.subtractQtyComments(comment.parentId)
     await this.repository.delete(commentId)
   }
 
@@ -45,10 +50,10 @@ export class CommentServiceImpl implements CommentService {
     return await this.repository.getByAuthorId(authorId)
   }
 
-  async getCommentsByPost (userId: string, postId: string): Promise<PostDTO[]> {
+  async getCommentsByPost (userId: string, postId: string, options: CursorPagination): Promise<ExtendedPostDTO[]> {
     const doesFollowExist = await followerService.doesFollowExist(userId, postId)
     const author = await userService.getUser(postId)
     if (!doesFollowExist && author.isPrivate) throw new NotFoundException('comment')
-    return await this.repository.getAllByPostId(postId)
+    return await this.repository.getAllByPostId(postId, options)
   }
 }
